Attack Combo
Attack combo animation implemented through code.

Code changed in Chip class and ChipAnimInstance class.
If want to only be in chip class, a timer would need to determine when
the user can click to activate next combo. For now the Montage is setup
and called in the ChipAnimInstance Class.

Currently does not play through the combo anims

+++ Source/TrashPanda/AI/Enemy.h
@@ -53,7 +53,7 @@ protected:
 		uint32 DamageReduction;
 
 	UPROPERTY(EditAnyWhere, Category = Stats)
-		uint32 Damage;
+		float Damage;
 
 	UPROPERTY(EditAnywhere, Category = AIBehavior)
 		uint32 MoveRadius;
